feat(stage): group persistent chat controls
